added pages and more components
