--- src/components/Lost/Lost.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Lost(){
-    return(
-        <div>
-            <h1></h1>
-        </div>
-    )
-}
