+++ src/index.js
@@ -2,12 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
-import Dashboard from "./pages/Dashboard/Dashboard.js";
 
 ReactDOM.render(
-  <React.StrictMode>
-    <Dashboard />
-  </React.StrictMode>,
+  <React.StrictMode></React.StrictMode>,
   document.getElementById("root")
 );
