--- src/pages/Homepage/Homepage.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Homepage(){
-    return(
-        <div>
-            <h1></h1>
-        </div>
-    )
-}
